refactor(cli): extract pane ops + send to pane_cli.py/send_cli.py (#495) (#531)
Phase 1 (Wave 1) of #495. Extracts pane operations (list, info, output,
kill, spawn, split, detach, jump, resize) into pane_cli.py and send/send-keys
into send_cli.py. Domain-private helpers (_install_pane_hooks, pane-0 state
probing, graceful-exit policy, worker-ready poll, .agentwire.yml type readers)
moved with the domain; shared helpers imported from core. Pure relocation,
CLI surface unchanged.

Built by dotdev.dev
